make ViewURLAsFile closeable and used the closeable try-catch rather than finalize on MorfXmlDatasetMatchers
